gcp/requirements.txt
Pillow==9.5.0
tensorflow==2.12.0
google-cloud-storage==2.8.0
python-multipart===0.0.6